refactor(cda): Refactorizamos endpoint de cda con cotroller
