+++ modules/cda/controller/CDAPatient.ts
@@ -14,6 +14,7 @@ import * as moment from 'moment';
 import { CDA as CDAConfig } from '../../../config.private';
 import { configuracionPrestacionModel } from './../../../core/term/schemas/configuracionPrestacion';
 import { Auth } from '../../../auth/auth.class';
+import { AndesDrive } from '@andes/drive';
 
 /**
  * Matcheamos los datos del paciente.
@@ -759,4 +760,45 @@ export function checkAndExtract(xmlDom) {
     }
 
     return metadata;
+}
+
+
+export async function getCda(params) {
+
+    const CDAFiles = makeFs();
+    const ObjectId = Types.ObjectId;
+    const cda = await CDAFiles.findOne({ _id: params.id });
+    if (cda) {
+        const adj = cda.metadata.adjuntos.find(_adj => {
+            return String(_adj.id) === params.realName;
+        });
+        if (adj && adj.adapter === 'drive') {
+
+            const fileDrive = await AndesDrive.find(ObjectId(params.realName));
+            if (fileDrive) {
+                const stream1 = await AndesDrive.read(fileDrive);
+                params.response.contentType(fileDrive.mimetype);
+                stream1.pipe(params.res);
+               return stream1;
+            }
+
+        } else {
+            const query = {
+                filename: params.name,
+               'metadata.cdaId': params.id
+            };
+            const file = await CDAFiles.findOne(query);
+            if (params.res.user && params.res.user.type === 'paciente-token' && String(file.metadata.paciente) !== String(params.res.user.pacientes[0].id)) {
+                throw 403
+            }
+            const stream1 = await CDAFiles.readFile({ _id: file._id });
+            params.res.contentType(file.contentType);
+            stream1.pipe(params.res);
+            return stream1;
+        }
+
+    } else {
+        return ('NO EXISTE EL ARCHIVO');
+    }
+
 }
